CoffeeCafe-POS/requirements.txt
# Requirements (requirements.txt)
# This file lists the dependencies required for the CoffeeCafe POS application.
# Ensure to keep this file updated with any new packages added to the project.

supabase==2.3.1
python-dotenv==1.0.0
customtkinter==5.2.1
passlib==1.7.4
pyjwt==2.8.0
pandas==2.0.3
python-escpos==3.0a8
black==23.7.0
ruff==0.0.280
matplotlib==3.7.1
numpy==1.24.3
pytest==7.4.0
pytest-mock==3.11.1